// src/app/modules/shared/models/addUser.interface.ts
// Generated by https://quicktype.io

export interface AddUser {
    address:  Address;
    email:    string;
    username: string;
    password: string;
    name:     Name;
    phone:    string;
}

export interface Address {
    geolocation: Geolocation;
    city:        string;
    street:      string;
    number:      number;
    zipcode:     string;
}

export interface Geolocation {
    lat:  string;
    long: string;
}

export interface Name {
    firstname: string;
    lastname:  string;
}