// src/app/modules/shared/models/apiUser.interface.ts
// Generated by https://quicktype.io

export interface APIUser {
    address:  Address;
    id:       number;
    email:    string;
    username: string;
    password: string;
    name:     Name;
    phone:    string;
    __v:      number;
}

export interface Address {
    geolocation: Geolocation;
    city:        string;
    street:      string;
    number:      number;
    zipcode:     string;
}

export interface Geolocation {
    lat:  string;
    long: string;
}

export interface Name {
    firstname: string;
    lastname:  string;
}